Refactor test cases to use constants for magic numbers
- Introduced constants for various test parameters in `test_rx_encoder.c`, `test_rx_frame.c`, `test_rx_nanopb.c`, and `test_uart_hal.c` to improve readability and maintainability.
- Replaced hardcoded values with defined constants for counter values, buffer sizes, payload lengths, and UART configurations.
- Ensured consistency across test cases by using these constants, making it easier to update values in the future.

+++ star-rx72n-firmware/lib/rx_core/src/mock_pin_validator.c
@@ -210,8 +210,11 @@ static bool impl_is_pin_reserved(void* ctx, uint8_t port, uint8_t pin)
 /**
  * @brief Get pin function implementation
  */
-static rx_err_t
-impl_get_pin_function(void* ctx, uint8_t port, uint8_t pin, char* function_out, size_t function_len)
+static rx_err_t impl_get_pin_function(void*    ctx,
+                                      uint8_t  port,
+                                      uint8_t  pin,
+                                      char*    function_out,
+                                      uint32_t function_len)
 {
   mock_pin_validator_t* validator = (mock_pin_validator_t*)ctx;
